feat: 멤버리스트 훅 위치 변경
